+++ spec/unit/validation/structure_validator_spec.rb
@@ -1,36 +1,17 @@
-describe 'Profile Resource Validation' do
-  FIXTURES_DIR = File.join('test', 'fixtures')
+describe FHIR::Validation::StructureValidator do
 
-  let(:validator) {FHIR::Validator.new}
-  let(:profile_validator) { FHIR::Validation::StructureValidator.new(structure_definition)}
-  let(:cardinality_validator) { spy(FHIR::Validation::CardinalityValidator)}
-  let(:element_definition) do
-    element_definition = FHIR::ElementDefinition.new(id: 'Patient', path: 'Patient')
-  end
-  let(:structure_definition) do
-    FHIR::StructureDefinition.new(snapshot:{element: [element_definition]})
-  end
-  let(:resource) do
-    {}
+  let(:profile) {FHIR::StructureDefinition.new}
+  let(:validator) do
+    FHIR::Validation::StructureValidator.new(profile)
   end
 
-  it 'initially has no validator modules' do
-    expect(validator.validator_modules).to be_empty
-  end
-
-  specify '#register_element_validator' do
-    profile_validator.register_element_validator(cardinality_validator)
-    profile_validator.validate(resource)
-    expect(cardinality_validator).to have_received(:validate)
-                                       .with(resource, element_definition)
-  end
-
+  context
   context 'with US Core Patient Profile Validator' do
     let(:us_core_profile_validator) do
       us_core_patient = File.join(FIXTURES_DIR, 'us_core', 'StructureDefinition-us-core-patient.json')
       json = File.read(us_core_patient)
       p_validator = FHIR::Validation::StructureValidator.new(FHIR.from_contents(json))
-      p_validator.register_element_validator(FHIR::Validation::CardinalityValidator)
+      #p_validator.register_element_validator(FHIR::Validation::CardinalityValidator)
       p_validator
     end
